Add guiding control methods and refactor into modules
Phase 2 implementation:
- Add start_guiding, stop_guiding, stop_capture, start_loop methods
- Add pause, resume, is_paused methods
- Add dither method with settling parameters
- Add CLI commands for all guiding control operations

Refactor library into focused modules:
- error.rs: Phd2Error and Result type
- rpc.rs: JSON RPC 2.0 types
- events.rs: PHD2 events and AppState
- config.rs: Configuration types
- types.rs: Rect, Profile
- client.rs: Phd2Client
- process.rs: Phd2ProcessManager

Update design documentation with module structure and
implementation status for phases 1-4.
