+++ services/phd2-guider/src/error.rs
@@ -1,0 +1,53 @@
+//! Error types for the PHD2 guider client
+
+/// Errors that can occur when interacting with PHD2
+#[derive(Debug, thiserror::Error)]
+pub enum Phd2Error {
+    #[error("Not connected to PHD2")]
+    NotConnected,
+
+    #[error("Connection failed: {0}")]
+    ConnectionFailed(String),
+
+    #[error("PHD2 not running")]
+    Phd2NotRunning,
+
+    #[error("Equipment not connected")]
+    EquipmentNotConnected,
+
+    #[error("Not calibrated")]
+    NotCalibrated,
+
+    #[error("Invalid state for operation: {0}")]
+    InvalidState(String),
+
+    #[error("RPC error: {code} - {message}")]
+    RpcError { code: i32, message: String },
+
+    #[error("Timeout: {0}")]
+    Timeout(String),
+
+    #[error("IO error: {0}")]
+    Io(#[from] std::io::Error),
+
+    #[error("JSON error: {0}")]
+    Json(#[from] serde_json::Error),
+
+    #[error("Failed to send message: {0}")]
+    SendError(String),
+
+    #[error("Failed to receive response")]
+    ReceiveError,
+
+    #[error("Failed to start PHD2 process: {0}")]
+    ProcessStartFailed(String),
+
+    #[error("PHD2 executable not found: {0}")]
+    ExecutableNotFound(String),
+
+    #[error("Process already running")]
+    ProcessAlreadyRunning,
+}
+
+/// Result type alias for PHD2 operations
+pub type Result<T> = std::result::Result<T, Phd2Error>;
